Run tests with and without Sphynx.

+++ .test_backend.sh
@@ -2,34 +2,44 @@
 
 cd `dirname $0`
 
-# Start Sphynx.
-TMP=$(mktemp -d)
-SPHYNX_PID_FILE=${TMP}/sphynx_pid
-export SPHYNX_PORT=$[ 9400 + RANDOM % 100 ]
-export SPHYNX_CERT_DIR=$TMP/sphynx_cert
-export ORDERED_SPHYNX_DATA_DIR=$TMP/ordered_sphynx_data
-export UNORDERED_SPHYNX_DATA_DIR=$TMP/unordered_sphynx_data
-if [ -f $SPHYNX_PID_FILE ]; then
-  kill `cat $SPHYNX_PID_FILE` || true
-fi
-mkdir -p ${SPHYNX_CERT_DIR}
-openssl req -x509 -sha256 -newkey rsa:4096 \
--keyout "${SPHYNX_CERT_DIR}/private-key.pem" \
--out "${SPHYNX_CERT_DIR}/cert.pem" -days 365 -nodes \
--subj "/C=/ST=/L=/O=Lynx Analytics/OU=Org/CN=localhost"
-stage/sphynx/go/bin/server -keydir=$SPHYNX_CERT_DIR &
-$(dirname $0)/tools/wait_for_port.sh $SPHYNX_PORT
-echo "Sphynx running on port $SPHYNX_PORT"
-echo $! > $SPHYNX_PID_FILE
+WITH_SPHYNX='false'
+while getopts 's' flag; do
+  case "${flag}" in
+    s) WITH_SPHYNX='true' ;;
+  esac
+done
+export WITH_SPHYNX=$WITH_SPHYNX
 
-function kill_sphynx {
-  echo "Shutting down Sphynx."
-  SPHYNX_PID=`cat ${SPHYNX_PID_FILE}`
-  kill $SPHYNX_PID
-  while kill -0 $SPHYNX_PID 2> /dev/null; do sleep 1; done
-  rm -rf "$TMP"
-}
-trap kill_sphynx EXIT ERR
+if $WITH_SPHYNX; then
+  # Start Sphynx.
+  TMP=$(mktemp -d)
+  SPHYNX_PID_FILE=${TMP}/sphynx_pid
+  export SPHYNX_PORT=$[ 9400 + RANDOM % 100 ]
+  export SPHYNX_CERT_DIR=$TMP/sphynx_cert
+  export ORDERED_SPHYNX_DATA_DIR=$TMP/ordered_sphynx_data
+  export UNORDERED_SPHYNX_DATA_DIR=$TMP/unordered_sphynx_data
+  if [ -f $SPHYNX_PID_FILE ]; then
+    kill `cat $SPHYNX_PID_FILE` || true
+  fi
+  mkdir -p ${SPHYNX_CERT_DIR}
+  openssl req -x509 -sha256 -newkey rsa:4096 \
+  -keyout "${SPHYNX_CERT_DIR}/private-key.pem" \
+  -out "${SPHYNX_CERT_DIR}/cert.pem" -days 365 -nodes \
+  -subj "/C=/ST=/L=/O=Lynx Analytics/OU=Org/CN=localhost"
+  stage/sphynx/go/bin/server -keydir=$SPHYNX_CERT_DIR &
+  $(dirname $0)/tools/wait_for_port.sh $SPHYNX_PORT
+  echo "Sphynx running on port $SPHYNX_PORT"
+  echo $! > $SPHYNX_PID_FILE
+
+  function kill_sphynx {
+    echo "Shutting down Sphynx."
+    SPHYNX_PID=`cat ${SPHYNX_PID_FILE}`
+    kill $SPHYNX_PID
+    while kill -0 $SPHYNX_PID 2> /dev/null; do sleep 1; done
+    rm -rf "$TMP"
+  }
+  trap kill_sphynx EXIT ERR
+fi
 
 mkdir -p logs
 rm -f logs/test-*
